Remove HTTP server usage of (*http.Request).PathValue and clarify requirements in godoc
