--- http/http_go_test.go
@@ -1,10 +1,0 @@
-// SPDX-FileCopyrightText: (C) 2024 Intel Corporation
-// SPDX-License-Identifier: Apache 2.0
-
-//go:build !tinygo
-
-package http_test
-
-import "net/http"
-
-func setPathValue(req *http.Request, name, value string) { req.SetPathValue(name, value) }
